intial commit

+++ src/main/java/edu/birzeit/ai_search/City.java
@@ -1,14 +1,14 @@
 package edu.birzeit.ai_search;
 
 public class City implements Comparable<City>{//Vertex
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private String name;
     private City parent;//
     double f;//function. f = g+h
     double travelledCost;//score of the cumulated actual distance
 
-    public City(int x, int y, String name){
+    public City(double x, double y, String name){
         this.x = x;
         this.y = y;
         this.name = name;
@@ -19,19 +19,19 @@ public class City implements Comparable<City>{//Vertex
         this.y = 0;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
